Move k/test/e2e_federation package to k/federation/test/e2e
