+++ federation/test/e2e/README.md
@@ -1,0 +1,1 @@
+See [e2e-tests](https://git.k8s.io/community/contributors/devel/e2e-tests.md#federation-e2e-tests)
