+++ federation/test/e2e/apiserver.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_federation
+package e2e
 
 import (
 	"fmt"
@@ -23,8 +23,8 @@ import (
 	. "github.com/onsi/gomega"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	fedframework "k8s.io/kubernetes/federation/test/e2e/framework"
 	"k8s.io/kubernetes/test/e2e/framework"
-	fedframework "k8s.io/kubernetes/test/e2e_federation/framework"
 )
 
 // Create/delete cluster api objects
