+++ federation/test/e2e/upgrades/BUILD
