--- test/e2e_federation/README.md
@@ -1,3 +1,0 @@
-See [e2e-tests](https://git.k8s.io/community/contributors/devel/e2e-tests.md#federation-e2e-tests)
-
-[![Analytics](https://kubernetes-site.appspot.com/UA-36037335-10/GitHub/test/e2e_federation//README.md?pixel)]()
